Update the ami to be compatible with EKS Kube 1.18 and above
The issue here was a mismatch between the ami we were using for the
self-managed nodes.
That ami worked OK up to EKS with Kube 1.18 and started failing with
Kube 1.19 and up with a cryptic error about issues with the CNI plugin
(the nodes were missing some config files although I suspect they were
missing some other things as well since replacing CNI with Calico did
not work either).

I have also updated the CF file to use the latest available Kube (1.21)
and updated the script installing a kubectl matching the version (it is
important to keep both in sync to prevent issues accordingly to the docs
 - and I verified myself lacking output from an old kubectl)